On windows, construct environment separately when storing to registry
We can't use the set of environment variable when setting local
valus and when storing to the registry.

This is because the new PATH value is constructed very differently
when updating the local shell compared to when updating the
registry.

Add test that verifies that no path elements are removed by
emsdk_env.  This includes not re-ordering or removing duplicates
that might exist already in the user's path.

Fixes: #634